+++ src/main/java/nextstep/courses/domain/session/Amount.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain.session.pay;
+package nextstep.courses.domain.session;
 
 import nextstep.courses.exception.NotMatchAmountException;
 import nextstep.payments.domain.Payment;
