+++ src/main/java/nextstep/courses/domain/session/Session.java
@@ -1,6 +1,7 @@
 package nextstep.courses.domain.session;
 
 import nextstep.courses.domain.session.coverimage.CoverImage;
+import nextstep.courses.domain.session.student.SessionStudents;
 import nextstep.users.domain.NsUser;
 
 public abstract class Session {
@@ -10,6 +11,8 @@ public abstract class Session {
     protected Status status;
     protected CoverImage coverImage;
     protected SessionStudents sessionStudents;
+
+    public Session() {}
 
     public Session(Long id, PayType payType, Status status, CoverImage coverImage) {
         this.id = id;
